Reorganized nodes
