+++ expressions.go
@@ -1,0 +1,5 @@
+package parser
+
+type expression interface {
+	evaluate() (interface{}, error)
+}
