+++ helpers.go
@@ -1,0 +1,19 @@
+package parser
+
+type node interface {
+	line() int
+	column() int
+}
+
+type columnDefinition struct {
+	identifier        string
+	nullable          bool
+	autoincrementable bool
+	defaultValue      interface{}
+	dataType
+}
+type whereClause struct {
+}
+
+type selectColumn interface {
+}
