+++ types.go
@@ -7,19 +7,6 @@ const (
 	floatSize   = 8
 )
 
-type node interface {
-	line() int
-	column() int
-}
-
-type expression interface {
-	evaluate() (interface{}, error)
-}
-
-type statement interface {
-	execute() error
-}
-
 type dataType interface {
 	size() int
 }
@@ -51,14 +38,4 @@ type floatType struct {
 
 func (t *floatType) size() int {
 	return floatSize
-}
-
-type columnDefinition struct {
-	identifier string
-	dataType   dataType
-}
-
-type createStatement struct {
-	identifier        string
-	columnDefinitions []columnDefinition
 }
